# apps/api/openapi.json
{"openapi":"3.0.0","paths":{"/secrets":{"get":{"operationId":"SecretsController_listSecrets","parameters":[],"responses":{"200":{"description":"List of secrets"}},"summary":"List all secrets (without actual values)","tags":["secrets"]},"post":{"operationId":"SecretsController_createSecret","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSecretDto"}}}},"responses":{"201":{"description":"Secret created"},"400":{"description":"Invalid input or duplicate name"}},"summary":"Create a new secret","tags":["secrets"]}},"/secrets/{id}":{"get":{"operationId":"SecretsController_getSecret","parameters":[{"name":"id","required":true,"in":"path","description":"Secret ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Secret details"},"404":{"description":"Secret not found"}},"summary":"Get a secret by ID (without actual value)","tags":["secrets"]},"put":{"operationId":"SecretsController_updateSecret","parameters":[{"name":"id","required":true,"in":"path","description":"Secret ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSecretDto"}}}},"responses":{"200":{"description":"Secret updated"},"404":{"description":"Secret not found"}},"summary":"Update a secret","tags":["secrets"]},"delete":{"operationId":"SecretsController_deleteSecret","parameters":[{"name":"id","required":true,"in":"path","description":"Secret ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Secret deleted"},"404":{"description":"Secret not found"}},"summary":"Delete a secret","tags":["secrets"]}},"/secrets/{id}/link/{workflowId}":{"post":{"operationId":"SecretsController_linkToWorkflow","parameters":[{"name":"id","required":true,"in":"path","description":"Secret ID","schema":{"type":"string"}},{"name":"workflowId","required":true,"in":"path","description":"Workflow ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Secret linked to workflow"}},"summary":"Link a secret to a workflow","tags":["secrets"]},"delete":{"operationId":"SecretsController_unlinkFromWorkflow","parameters":[{"name":"id","required":true,"in":"path","description":"Secret ID","schema":{"type":"string"}},{"name":"workflowId","required":true,"in":"path","description":"Workflow ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Secret unlinked from workflow"}},"summary":"Unlink a secret from a workflow","tags":["secrets"]}},"/api/variables":{"get":{"operationId":"VariablesController_getAll","parameters":[{"name":"scope","required":true,"in":"query","schema":{"type":"string"}},{"name":"workflowId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Variables"]},"post":{"operationId":"VariablesController_create","parameters":[],"responses":{"201":{"description":""}},"tags":["Variables"]}},"/api/variables/global":{"get":{"operationId":"VariablesController_getGlobal","parameters":[],"responses":{"200":{"description":""}},"tags":["Variables"]}},"/api/variables/definitions":{"get":{"operationId":"VariablesController_getWorkflowDefinitions","parameters":[],"responses":{"200":{"description":""}},"tags":["Variables"]}},"/api/variables/workflow/{workflowId}":{"get":{"operationId":"VariablesController_getByWorkflow","parameters":[{"name":"workflowId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Variables"]}},"/api/variables/{id}":{"get":{"operationId":"VariablesController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Variables"]},"put":{"operationId":"VariablesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Variables"]},"delete":{"operationId":"VariablesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Variables"]}},"/scrape/stop":{"post":{"description":"Stops the currently running scrape operation","operationId":"ScrapeController_stopScrape","parameters":[],"responses":{"200":{"description":"Scrape stopped successfully"},"404":{"description":"No running scrape found"}},"summary":"Stop running scrape","tags":["scrapes"]}},"/scrape/{scrapeId}":{"get":{"description":"Starts a scraping job with the specified configuration ID","operationId":"ScrapeController_scrape","parameters":[{"name":"scrapeId","required":true,"in":"path","description":"The ID of the scrape configuration to execute","schema":{"example":"amazon-orders","type":"string"}}],"responses":{"200":{"description":"Scraping started successfully","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"running"},"scrapeId":{"type":"string","example":"amazon-orders"}}}}}},"404":{"description":"Scrape configuration not found"}},"summary":"Start scraping","tags":["scrapes"]}},"/scrapes":{"get":{"operationId":"ScrapeUIController_getScrapes","parameters":[],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/scrapes/{id}":{"get":{"operationId":"ScrapeUIController_getScrapeById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/run/{scrapeId}":{"post":{"operationId":"ScrapeUIController_runScrape","parameters":[],"responses":{"201":{"description":""}},"tags":["ScrapeUI"]}},"/events":{"get":{"operationId":"ScrapeUIController_events","parameters":[],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/events/status":{"get":{"operationId":"ScrapeUIController_getEventsStatus","parameters":[],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/events/ping":{"post":{"operationId":"ScrapeUIController_pingEvents","parameters":[],"responses":{"201":{"description":""}},"tags":["ScrapeUI"]}},"/logs":{"get":{"operationId":"ScrapeUIController_getLogs","parameters":[],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/logs/clear":{"post":{"operationId":"ScrapeUIController_clearLogs","parameters":[],"responses":{"201":{"description":""}},"tags":["ScrapeUI"]}},"/otp/{requestId}":{"post":{"operationId":"ScrapeUIController_submitOtp","parameters":[{"name":"requestId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["ScrapeUI"]}},"/runs":{"get":{"operationId":"ScrapeUIController_getRuns","parameters":[{"name":"scrapeId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/runs/{runId}":{"get":{"operationId":"ScrapeUIController_getRunById","parameters":[{"name":"runId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]},"delete":{"operationId":"ScrapeUIController_deleteRun","parameters":[{"name":"runId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/scrapes/{scrapeId}/runs":{"delete":{"operationId":"ScrapeUIController_deleteRunsByScrapeId","parameters":[{"name":"scrapeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/runs/cleanup":{"post":{"operationId":"ScrapeUIController_cleanupRuns","parameters":[],"responses":{"201":{"description":""}},"tags":["ScrapeUI"]}},"/scrapes/{scrapeId}/data":{"get":{"operationId":"ScrapeUIController_getScrapeData","parameters":[{"name":"scrapeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/runs/{runId}/data":{"get":{"operationId":"ScrapeUIController_getRunData","parameters":[{"name":"runId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/scrapes/{scrapeId}/data/all":{"get":{"operationId":"ScrapeUIController_getAllScrapeData","parameters":[{"name":"scrapeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/scrapes/{scrapeId}/schedule":{"get":{"operationId":"ScrapeUIController_getSchedule","parameters":[{"name":"scrapeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]},"put":{"operationId":"ScrapeUIController_updateSchedule","parameters":[{"name":"scrapeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/schedules":{"get":{"operationId":"ScrapeUIController_getAllSchedules","parameters":[],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/scheduler/status":{"get":{"operationId":"ScrapeUIController_getSchedulerStatus","parameters":[],"responses":{"200":{"description":""}},"tags":["ScrapeUI"]}},"/api/actions/metadata":{"get":{"operationId":"ActionsController_getActionsMetadata","parameters":[],"responses":{"200":{"description":""}},"tags":["Actions"]}},"/api/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}},"/api/health/live":{"get":{"operationId":"HealthController_liveness","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}},"/api/health/ready":{"get":{"operationId":"HealthController_readiness","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}}},"info":{"title":"Scrape Dojo API","description":"API für modulares Web-Scraping mit Puppeteer","version":"1.0","contact":{}},"tags":[{"name":"scrapes","description":"Scrape-Definitionen und Ausführung"},{"name":"secrets","description":"Secrets Management"},{"name":"logs","description":"Logging und Events"}],"servers":[{"url":"http://localhost:3000","description":"Development Server"}],"components":{"schemas":{"CreateSecretDto":{"type":"object","properties":{}},"UpdateSecretDto":{"type":"object","properties":{}}}}}